% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pathways.R
\name{analyze_pathways}
\alias{analyze_pathways}
\title{Analyze Pathways}
\usage{
analyze_pathways(genes, pathways_file, gene_number_cutoff, mode,
  sample_size, num_cores)
}
\arguments{
\item{genes}{Genes from parse_SNP()}

\item{pathways_file}{A file containing the pathway IDs, their names, and the genes in the pathway}

\item{gene_number_cutoff}{A cut-off for the minimum number of genes in a pathway}

\item{mode}{increasing/decreasing}

\item{sample_size}{How many times to sample the effects data during random sampling}

\item{num_cores}{The number of cores to use in parallelizing PAST}
}
\value{
Rugplots data
}
\description{
Analyze Pathways
}
\examples{
demo_pathways_file = system.file("extdata", "pathways.txt.xz", package = "PAST", mustWork = TRUE)
data(genes)
rugplots_data_susceptible <- analyze_pathways(genes, demo_pathways_file, 5, "increasing", 1000, 2)
}
