% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Coerce to a Data Frame}
\usage{
as.data.frame(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[base:as.data.frame]{base::as.data.frame()}}}
}
\description{
Drop in replacements for \code{\link[base:as.data.frame]{base::as.data.frame()}} with \code{stringsAsFactos = TRUE}
by default.
}
\details{
In R version 4.0.0 the \code{stringsAsFactors} default was changed to \code{FALSE}. In
addition, using \code{options(stringsAsFactors = TRUE)} now throws a nasty warning.
If your code depended on the old \code{stringsAsFactors = TRUE} default and you
want to upgrade your R version then you either have to set the argument explicitly
for every function call or you add \code{library(past)} at the top of your program.
Presumably, the latter is an easier and less time consuming option.
}
\examples{
df <- as.data.frame(letters)
is.factor(df$letters)

}
\author{
Thomas Neitmann
}
