% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_SNPs_to_genes.R
\name{assign_chunk}
\alias{assign_chunk}
\title{Assign SNPs in a chunk to genes}
\usage{
assign_chunk(gff, chunk, window)
}
\arguments{
\item{gff}{The GFF data for the chromosome being parsed}

\item{chunk}{The dataframe containing SNP data}

\item{window}{The search window around the SNPs}
}
\value{
tagSNPs labeled with gene names
}
\description{
Assign SNPs in a chunk to genes
}
