% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_SNPs_to_genes.R
\name{assign_SNPs_to_genes}
\alias{assign_SNPs_to_genes}
\title{Assign SNPs to genes}
\usage{
assign_SNPs_to_genes(
  gwas_data,
  LD,
  GFF_file,
  filter_type = "gene",
  window,
  r_squared_cutoff,
  name_attribute,
  debug = FALSE
)
}
\arguments{
\item{gwas_data}{GWAS data from load_GWAS_data()}

\item{LD}{linkage disequilibrium data from load_LD()}

\item{GFF_file}{the path to a GFF-formatted of annotations.}

\item{filter_type}{keep only this type of feature from the GFF}

\item{window}{the search window for genes around the SNP}

\item{r_squared_cutoff}{the R^2 value used to determine SNP significance}

\item{name_attribute}{the field in the 9th column of the GFF file that
contains the desired gene name (for example, the 9th column might look like
ID=NNN;Name=NNN;description=NNN - to use the name in the Name field, provide
"Name")}

\item{debug}{return all SNPs associated with a gene}
}
\value{
a data.table of SNP-gene assignments, reduced to the SNP that best
represents all SNPs assigned to the same gene for each gene
}
\description{
Assign SNPs to genes
}
\examples{
example("load_GWAS_data")
example("load_LD")
GFF = system.file("extdata", "genes.gff.gz", package = "PAST2",
  mustWork = TRUE)
genes <- assign_SNPs_to_genes(gwas_data, LD, GFF, "gene", 1000, 0.8, "Name")
}
