% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{data.frame}
\alias{data.frame}
\title{Data Frames}
\usage{
data.frame(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[base:data.frame]{base::data.frame()}}}
}
\description{
Drop in replacements for \code{\link[base:data.frame]{base::data.frame()}} with \code{stringsAsFactos = TRUE}
by default.
}
\details{
In R version 4.0.0 the \code{stringsAsFactors} default was changed to \code{FALSE}. In
addition, using \code{options(stringsAsFactors = TRUE)} now throws a nasty warning.
If your code depended on the old \code{stringsAsFactors = TRUE} default you
want to upgrade your R version then you either have to set the argument explicitly
for every function call or you add \code{library(past)} at the top of your program.
Presumably, the latter is an easier and less time consuming option.
}
\examples{
df <- data.frame(char = letters, num = 1:26)
is.factor(df$char)

}
\author{
Thomas Neitmann
}
