% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_SNP.R
\name{determine_linkage}
\alias{determine_linkage}
\title{Determine Linkage}
\usage{
determine_linkage(chunk, r_squared_cutoff)
}
\arguments{
\item{chunk}{A chunk of data to be processed}

\item{r_squared_cutoff}{The R^2 value to check against}
}
\value{
Either the first unlinked SNP or a set of linked SNPs
}
\description{
Determine Linkage
}
