% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_SNP.R
\name{find_genes}
\alias{find_genes}
\title{Find genes within a window of a SNP}
\usage{
find_genes(gff, snp_df, window)
}
\arguments{
\item{gff}{The GFF data for the chromosome being parsed}

\item{snp_df}{The dataframe containing SNP data}

\item{window}{The search window around the SNPs}
}
\value{
SNPs labeled with gene names
}
\description{
Find genes within a window of a SNP
}
