% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_SNP.R
\name{find_representative_SNP}
\alias{find_representative_SNP}
\title{Find representative SNP for a chunk of SNPs}
\usage{
find_representative_SNP(chunk, r_squared_cutoff)
}
\arguments{
\item{chunk}{A chunk of data to parse}

\item{r_squared_cutoff}{The R^2 value to check against when counting SNPs}
}
\value{
A single SNP representing the whole chunk
}
\description{
Find representative SNP for a chunk of SNPs
}
