% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_SNPs_to_genes.R
\name{find_representative_SNP_gene_pairing}
\alias{find_representative_SNP_gene_pairing}
\title{Find the SNP-gene assignment that represents SNPs assigned to a gene}
\usage{
find_representative_SNP_gene_pairing(chunk)
}
\arguments{
\item{chunk}{A chunk of gene assignments}
}
\value{
A single SNP-gene assignment representing all SNPS assigned to the
  same gene
to a gene
}
\description{
Find the SNP-gene assignment that represents SNPs assigned to a gene
}
