% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{genes}
\alias{genes}
\title{Genes data}
\format{A data frame with 183 rows and 5 variables:
\describe{
  \item{Chromosome}{The chromosome on which the SNP associated with the gene is found.}
  \item{Position}{The physical location, in base pairs from the start of the chromosome, where the SNP associated with the gene is found. This may vary depending on the reference sequence version which was used to align the SNPs.}
  \item{Gene}{The name of the gene}
  \item{Effect}{The phenotypic effect that the allele had on the trait. The strongest effect is assigned to the gene.}
  \item{P_value}{The p-value from the T-test of each individual marker-trait association for the SNP associated with the gene.}
}}
\usage{
genes
}
\description{
A dataset containing the assigned genes
}
\keyword{datasets}
