% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_file}
\alias{load_file}
\title{Load GWAS data}
\usage{
load_file(file, columns, file_has_header = TRUE)
}
\arguments{
\item{file}{the file to read}

\item{columns}{a named list of the columns describing what they are and the
user-provided name}

\item{file_has_header}{a boolean indicating whether the file has a header
or not}
}
\value{
file with columns specified by user named according to names of
columns list
}
\description{
Load GWAS data
}
