% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_LD.R
\name{load_LD}
\alias{load_LD}
\title{Load Linkage Disequilibrium}
\usage{
load_LD(
  LD_file,
  locus = "Locus1",
  position1 = "Position1",
  site1 = "Site1",
  position2 = "Position2",
  site2 = "Site2",
  distance = "Dist_bp",
  r_squared = "R^2"
)
}
\arguments{
\item{LD_file}{the file containing linkage disequilibrium data}

\item{locus}{the name of the column containing the locus or chromosome}

\item{position1}{the name of the column containing the first SNP's position}

\item{site1}{the name of the column containing the first SNP's site}

\item{position2}{the name of the column containing second SNP's position}

\item{site2}{the name of the column containing the second SNP's site}

\item{distance}{the name of the column containing distance between the SNPs}

\item{r_squared}{the name of the column containing R^2}
}
\value{
the linkage disequilibrium data in a data.table
}
\description{
Load Linkage Disequilibrium
}
\examples{
LD_file = system.file("extdata","LD.txt.gz", package = "PAST2",
mustWork = TRUE)
LD <- load_LD(LD_file, r_squared="R.2")
}
