% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GWAS_data.R
\name{load_TASSEL}
\alias{load_TASSEL}
\title{Load TASSEL data}
\usage{
load_TASSEL(
  association_file,
  effects_file,
  trait = "Trait",
  marker = "Marker",
  locus = "Locus",
  site = "Site",
  p = "p",
  marker_R2 = "marker_R2",
  effect = "Effect",
  effects_trait = "Trait",
  effects_marker = "Marker",
  effects_locus = "Locus",
  effects_site = "Site",
  update_progress = NULL
)
}
\arguments{
\item{association_file}{The association file}

\item{effects_file}{The effects file}

\item{trait}{The names of the trait column}

\item{marker}{The names of the marker column}

\item{locus}{The names of the locus column}

\item{site}{The names of the site column}

\item{p}{The names of the p column}

\item{marker_R2}{The names of the marker_R2 column}

\item{effect}{The names of the effect column}

\item{update_progress}{an optional function for use with shiny that updates the user on progress}
}
\value{
The association data and the effects data merged into a dataframe
  with one row for each SNP
}
\description{
Load TASSEL data
}
