% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_effects_stats.R
\name{merge_data}
\alias{merge_data}
\title{Merge Data}
\usage{
merge_data(association_file, effects_file)
}
\arguments{
\item{association_file}{The association file}

\item{effects_file}{The effects file}
}
\value{
The association data and the effects data merged into a dataframe
  with one row for each SNP
}
\description{
Merge Data
}
\examples{
demo_association_file = system.file("extdata",
"association.txt.xz", package = "PAST", mustWork = TRUE)
demo_effects_file = system.file("extdata",
"effects.txt.xz", package = "PAST", mustWork = TRUE)
merged_data <- merge_data(demo_association_file, demo_effects_file)
}
