% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_SNP.R
\name{parse_chunk}
\alias{parse_chunk}
\title{Parse blocked SNPs into a single representative SNP}
\usage{
parse_chunk(block, r_squared_cutoff)
}
\arguments{
\item{block}{A block of data to parse}

\item{r_squared_cutoff}{The R^2 value to check against when counting SNPs}
}
\value{
A single SNP representing the whole block
}
\description{
Parse blocked SNPs into a single representative SNP
}
