% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_LD.R
\name{parse_LD}
\alias{parse_LD}
\title{Parse Linkage Disequilibrium}
\usage{
parse_LD(LD_file)
}
\arguments{
\item{LD_file}{The file containing linkage disequilibrium data (accepts plain-text or gzipped plain-text)}
}
\value{
The linkage disequilibrium data in a list containing dataframes for each chromosome.
}
\description{
Parse Linkage Disequilibrium
}
\examples{
demo_linkage_disequilibrium_file = system.file("extdata", 
"LD.txt.xz", package = "PAST", mustWork = TRUE)
LD <- parse_LD(demo_linkage_disequilibrium_file)
}
