% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_SNP.R
\name{parse_temp}
\alias{parse_temp}
\title{Parse Chunks of Data for Significant SNPs}
\usage{
parse_temp(chunk, r_squared_cutoff)
}
\arguments{
\item{chunk}{A chunk of data to be processed}

\item{r_squared_cutoff}{The R^2 value to check against}
}
\value{
Either the first unlinked SNP or a set of linked SNPs
}
\description{
Parse Chunks of Data for Significant SNPs
}
