% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{plot_enrichment_data}
\alias{plot_enrichment_data}
\title{Create plots for pathways based on user-provided filtering strategy.}
\usage{
plot_enrichment_data(
  enrichment_data,
  significance_measure = "q-value",
  significance_cutoff = 1,
  output_directory,
  theme = "dark"
)
}
\arguments{
\item{enrichment_data}{the data to be plotted (returned from
find_pathway_significance())}

\item{significance_measure}{the parameter to be used for filtering}

\item{significance_cutoff}{the cut-off value of the significance parameter}

\item{output_directory}{a directory to save results in; will be created if it}

\item{theme}{set image to light or dark theme
doesn't exist}
}
\value{
no return
}
\description{
Create plots for pathways based on user-provided filtering strategy.
}
\examples{
example("find_pathway_significance")
plot_enrichment_data(enrichment_data, "q-value", "0.05", tempdir())
}
