% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rugplots.R
\name{plot_pathways}
\alias{plot_pathways}
\title{Plot Pathways}
\usage{
plot_pathways(rugplots_data, filter_type, filter_parameter, mode,
  output_directory)
}
\arguments{
\item{rugplots_data}{The data to be plotted (returned from parse_pathways())}

\item{filter_type}{The parameter to be used for filtering}

\item{filter_parameter}{The cut-off value of the filtering parameter}

\item{mode}{The mode used to create the data (increasing/decreasing)}

\item{output_directory}{An existing directory to save results in}
}
\description{
Plot Pathways
}
\examples{
data(rugplots_data)
plot_pathways(rugplots_data, "pvalue", "0.03", "decreasing", "results")
}
