% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.table}
\alias{read.table}
\alias{read.csv}
\alias{read.csv2}
\alias{read.delim}
\alias{read.delim2}
\title{Data Input}
\usage{
read.table(...)

read.csv(...)

read.csv2(...)

read.delim(...)

read.delim2(...)
}
\arguments{
\item{...}{Arguments passed on the the respective function in \code{\link{utils}}}
}
\description{
Drop in replacements for the \verb{utils::read_*()} functions with \code{stringsAsFactos = TRUE}
by default.
}
\details{
In R version 4.0.0 the \code{stringsAsFactors} default was changed to \code{FALSE}. In
addition, using \code{options(stringsAsFactors = TRUE)} now throws a nasty warning.
If your code depended on the old \code{stringsAsFactors = TRUE} default and you
want to upgrade your R version then you either have to set the argument explicitly
for every function call or you add \code{library(past)} at the top of your program.
Presumably, the latter is an easier and less time consuming option.
}
\examples{
file <- tempfile()
write.csv(iris, file = file, row.names = FALSE)
iris2 <- read.csv(file)
is.factor(iris2$Species)

}
\author{
Thomas Neitmann
}
