% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_SNP.R
\name{read_gff}
\alias{read_gff}
\title{Read a GFF file}
\usage{
read_gff(gff_file)
}
\arguments{
\item{gff_file}{The path to the GFF file}
}
\value{
A dataframe containing the relevant GFF data - chromosome, gene name, start and stop locations
}
\description{
Read a GFF file
}
