% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GWAS_data.R
\name{read_GWAS_data}
\alias{read_GWAS_data}
\title{Load GWAS data}
\usage{
read_GWAS_data(association_file, effects_file, stats_columns = c("Trait",
  "Marker", "Locus", "Site", "F", "p", "marker_R2"),
  effects_columns = c("Trait", "Marker", "Locus", "Site", "Effect"))
}
\arguments{
\item{association_file}{The association file}

\item{effects_file}{The effects file}

\item{type}{The type of analysis from TASSEL (mlm or glm)}
}
\value{
The association data and the effects data merged into a dataframe
  with one row for each SNP
}
\description{
Load GWAS data
}
\examples{
demo_association_file = system.file("extdata",
"association.txt.xz", package = "PAST", mustWork = TRUE)
demo_effects_file = system.file("extdata",
"effects.txt.xz", package = "PAST", mustWork = TRUE)
gwas_data <- read_GWAS_data(demo_association_file, demo_effects_file)
}
