% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_SNP.R
\name{select_gene_from_block}
\alias{select_gene_from_block}
\title{Select the SNP with the largest effect
from all SNPs assigned to a gene}
\usage{
select_gene_from_block(untagged_genes)
}
\arguments{
\item{untagged_genes}{A chunk of untagged genes}
}
\value{
A single SNP representing all SNPS assigned
to a gene
}
\description{
Select the SNP with the largest effect
from all SNPs assigned to a gene
}
