% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GWAS_data.R
\name{load_GWAS_data}
\alias{load_GWAS_data}
\title{Load GWAS data}
\usage{
load_GWAS_data(
  first_file,
  second_file = NULL,
  input_type,
  mode,
  trait,
  marker,
  locus,
  site,
  p,
  marker_R2,
  effect,
  effects_trait = NULL,
  effects_marker = NULL,
  effects_locus = NULL,
  effects_site = NULL,
  effects_allele = NULL
)
}
\arguments{
\item{first_file}{a file containing combined GWAS data or containing
association data}

\item{second_file}{a file containing effects data; only use when first_file
contains association data}

\item{input_type}{one of "single", "two", or "TASSEL"}

\item{mode}{homozygous or heterozygous; retains either only homozygous data
or homozygous and heterozygous data}

\item{trait}{the name of the column containing the trait}

\item{marker}{the name of the column containing the marker}

\item{locus}{the name of the column containing the locus or chromosome}

\item{site}{the name of the column containing the site or position}

\item{p}{the name of the column containing the p-value}

\item{marker_R2}{the name of the column containing the p-value}

\item{effect}{the name of the column containing the effect}

\item{effects_trait}{the name of the column containing the trait in the
effects file; only used when input_type is "two" or "TASSEL"}

\item{effects_marker}{the name of the column containing the marker in the
effects file; only used when input_type is "two" or "TASSEL"}

\item{effects_locus}{the name of the column containing the locus or
chromosome in the effects file; only used when input_type is "two" or
"TASSEL"}

\item{effects_site}{the name of the column containing the site or position
in the effects file; only used when input_type is "two" or "TASSEL"}

\item{effects_allele}{the name of the column containing the allele in the
effects file; only used when input_type is "TASSEL" and mode is
"increasing" or "decreasing"}
}
\value{
GWAS data with a marker created from locus/chromosome and
site/position, the original marker, locus/chromosome, site/position, p-value,
marker R^2, and effect for every marker in the input files
}
\description{
Load GWAS data
}
\examples{
association <- system.file("extdata", "association.txt.gz",
    package = "PAST2",
    mustWork = TRUE
)
effects <- system.file("extdata", "effects.txt.gz",
    package = "PAST2",
    mustWork = TRUE
)
single <- system.file("extdata", "single_gwas.txt.gz",
    package = "PAST2",
    mustWork = TRUE
)
effects_single <- system.file("extdata", "effects.single_line.txt.gz",
    package = "PAST2", mustWork = TRUE
)
gwas_data <- load_GWAS_data(
    single,
    NULL,
    "single",
    "homozygous",
    trait = "Trait",
    marker = "Marker",
    locus = "Locus",
    site = "Site",
    p = "p",
    marker_R2 = "marker_R2",
    effect = "Effect"
)
gwas_data <- load_GWAS_data(
    association,
    effects_single,
    "two",
    "homozygous",
    trait = "Trait",
    marker = "Marker",
    locus = "Locus",
    site = "Site",
    p = "p",
    marker_R2 = "marker_R2",
    effect = "Effect"
)
gwas_data <- load_GWAS_data(
    association,
    effects,
    "TASSEL",
    "homozygous",
    trait = "Trait",
    marker = "Marker",
    locus = "Locus",
    site = "Site",
    p = "p",
    marker_R2 = "marker_R2",
    effect = "Effect"
)
}
