% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GWAS_data.R
\name{load_GWAS_data}
\alias{load_GWAS_data}
\title{Load GWAS data}
\usage{
load_GWAS_data(
  files,
  input_type,
  trait = "Trait",
  marker = "Marker",
  locus = "Locus",
  site = "Site",
  p = "p",
  marker_R2 = "marker_R2",
  effect = "Effect",
  effects_trait = "Trait",
  effects_marker = "Marker",
  effects_locus = "Locus",
  effects_site = "Site",
  update_progress = NULL
)
}
\arguments{
\item{files}{a vector of files - either a single GWAS file, two files with
separate association and effects data with one marker per line, or two files
from the output of TASSEL}

\item{input_type}{one of "single", "two", or "TASSEL"}

\item{trait}{the name of the column containing the trait}

\item{marker}{the name of the column containing the marker}

\item{locus}{the name of the column containing the locus or chromosome}

\item{site}{the name of the column containing the site or position}

\item{p}{the name of the column containing the p-value}

\item{marker_R2}{the name of the column containing the p-value}

\item{effect}{the name of the column containing the effect}

\item{effects_trait}{the name of the column containing the trait in the 
effects file; only used when input_type is "two" or "TASSEL"}

\item{effects_marker}{the name of the column containing the marker in the 
effects file; only used when input_type is "two" or "TASSEL"}

\item{effects_locus}{the name of the column containing the locus or 
chromosome in the effects file; only used when input_type is "two" or 
"TASSEL"}

\item{effects_site}{the name of the column containing the site or position
in the effects file; only used when input_type is "two" or "TASSEL"}

\item{update_progress}{used by the PAST Shiny app to report progress}
}
\value{
GWAS data with a marker created from locus/chromosome and 
site/position, the original marker, locus/chromosome, site/position, p-value,
marker R^2, and effect for every marker in the input files
}
\description{
Load GWAS data
}
\examples{
association = system.file("extdata", "association.txt.gz", package = "PAST", 
  mustWork = TRUE)
effects = system.file("extdata", "effects.txt.gz", package = "PAST", 
  mustWork = TRUE)
single = system.file("extdata", "single_gwas.txt.gz", package = "PAST", 
  mustWork = TRUE)
effects_single = system.file("extdata", "effects.single_line.txt.gz", 
  package = "PAST", mustWork = TRUE)
  
one_file = c(single_file)
two_file = c(association_file, effects_single_file)
tassel = c(association_file, effects_file)
gwas_data <- load_GWAS_data(one_file, "single")
gwas_data <- load_GWAS_data(two_file, "two")
gwas_data <- load_GWAS_data(tassel, "TASSEL")
}
