% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_SNP.R
\name{parse_SNP}
\alias{parse_SNP}
\title{Parse SNP data and return genes}
\usage{
parse_SNP(all_data, LD, gff_file, window, r_squared_cutoff, num_cores)
}
\arguments{
\item{all_data}{Merged association and effects data from merge_data()}

\item{LD}{Linkage disequilibrium data from parse_LD()}

\item{gff_file}{The path to a GFF file (accepts plain-text or gzipped plain-text)}

\item{window}{The search window for genes around the SNP}

\item{r_squared_cutoff}{The R^2 value used to determine SNP significance}

\item{num_cores}{The number of cores to use in parallelizing PAST}
}
\value{
A dataframe of genes from the SNP data
}
\description{
Parse SNP data and return genes
}
\examples{
demo_association_file = system.file("extdata", "association.txt.xz",
  package = "PAST", mustWork = TRUE)
demo_effects_file = system.file("extdata", "effects.txt.xz",
  package = "PAST", mustWork = TRUE)
gwas_data <- load_GWAS_data(demo_association_file, demo_effects_file)
demo_linkage_disequilibrium_file = system.file("extdata","LD.txt.xz",
package = "PAST", mustWork = TRUE)
LD <- load_LD(demo_linkage_disequilibrium_file)
demo_genes_file = system.file("extdata", "genes.gff.xz",
  package = "PAST", mustWork = TRUE)
genes <-parse_SNP(gwas_data, LD, demo_genes_file, 1000, 0.8, 2)
}
