\name{parse_SNP}
\alias{parse_SNP}
\title{parses LD file by filtering for SNPs that meet a cutoff value and detecting linkage}
\usage{
  parse_SNP(all_data, LD)
}
\description{
  Parses LD file by filtering for SNPs that meet a cutoff value and detecting linkage
}
\examples{
  parse_SNP(all_data, LD)
}
