% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{rugplots_data}
\alias{rugplots_data}
\title{A dataframe containing the necessary information
to create the rugplots in the next step}
\format{An object of class \code{data.frame} with 17 rows and 9 columns.}
\usage{
rugplots_data
}
\description{
A dataframe containing the necessary information
to create the rugplots in the next step
}
\keyword{datasets}
